fix ipq_id_t on 'real' kernel+userspace 64bit archs (Ryan Veety)

+++ include/libipq/libipq.h
@@ -35,7 +35,7 @@
 typedef u_int64_t ipq_id_t;
 #else
 #include <linux/netfilter_ipv4/ip_queue.h>
-typedef u_int32_t ipq_id_t;
+typedef unsigned long ipq_id_t;
 #endif
 
 #ifdef DEBUG_LIBIPQ
